+++ projects/pri-ng-scrollbar/src/lib/pri.scrollbar.component.ts
@@ -56,6 +56,10 @@ export class PriScrollbarComponent implements AfterViewInit, OnDestroy, AfterVie
    * smaller, but we will ocver scrollbar with size smaller than this value)
    */
   private readonly ASSUME_SCROLLBAR_SIZE = 20;
+  /**
+   * when overflow x is not hidden we use a smaller scrollbar size not to shrink content (this is more or less a workaround)
+   */
+  private readonly ASSUME_SCROLLBAR_SIZE_OX = 17;
   /** check if size changed (we are only reacting to changes if its more than 2px)
    * because we have set a width in percentage, which can cause rounding problems especially with different zoom factors
    * to get rid of the rounding differences, and prevent to recalculate bec. of those, we use 2px as threshold */
@@ -369,7 +373,11 @@ export class PriScrollbarComponent implements AfterViewInit, OnDestroy, AfterVie
       // based on the "ASSUME_SCROLLBAR_SIZE" description, we always take hide scrollbars up to 20px with a static value
       // this prevent browser rounding problems, when zoom factor or percent width/height is set
       if (nativeScrollbarSize > 0) {
-        nativeScrollbarSize = nativeScrollbarSize <= this.ASSUME_SCROLLBAR_SIZE ? this.ASSUME_SCROLLBAR_SIZE : nativeScrollbarSize;
+        // we will always use the assumed scrollbar size
+        nativeScrollbarSize = this.overflowX !== PriScrollbarOverflowTypes.hidden
+          ? this.ASSUME_SCROLLBAR_SIZE_OX
+          : this.ASSUME_SCROLLBAR_SIZE;
+        // nativeScrollbarSize = nativeScrollbarSize <= this.ASSUME_SCROLLBAR_SIZE ? this.ASSUME_SCROLLBAR_SIZE : nativeScrollbarSize;
       }
       // show vertical scroll bar
       // if native scrollbar size is 0 its a floatingScrollW. and we cant use pri-scrollbar