+++ src/app/+development/development.second.component.ts
@@ -1,0 +1,11 @@
+import {ChangeDetectionStrategy, Component} from '@angular/core';
+
+@Component({
+  selector: 'app-development-second',
+  templateUrl: 'development.second.component.html',
+  styleUrls: ['development.second.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+
+export class DevelopmentSecondComponent {
+}